remove rolling stuff in proba
