+++ river/proba/__init__.py
@@ -2,7 +2,5 @@
 from . import base
 from .gaussian import Gaussian
 from .multinomial import Multinomial
-from .rolling import Rolling
-from .time_rolling import TimeRolling
 
-__all__ = ["base", "Gaussian", "Multinomial", "Rolling", "TimeRolling"]
+__all__ = ["base", "Gaussian", "Multinomial"]
